% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossrunAnalyzer-package.R
\docType{package}
\name{lossrunAnalyzer-package}
\alias{lossrunAnalyzer}
\alias{lossrunAnalyzer-package}
\title{lossrunAnalyzer: R Package and Shiny App to Analyze Insurance Lossruns}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/lossrunAnalyzer}
  \item Report bugs at \url{https://github.com/jimbrig2011/lossrunAnalyzer/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
